+++ src/SoftDeleteable/Event/PreSoftDeleteEventArgs.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Doctrine Behavioral Extensions package.
+ * (c) Gediminas Morkevicius <gediminas.morkevicius@gmail.com> http://www.gediminasm.org
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Gedmo\SoftDeleteable\Event;
+
+use Doctrine\Persistence\Event\LifecycleEventArgs;
+use Doctrine\Persistence\ObjectManager;
+
+/**
+ * @template TObjectManager of ObjectManager
+ *
+ * @template-extends LifecycleEventArgs<TObjectManager>
+ */
+final class PreSoftDeleteEventArgs extends LifecycleEventArgs
+{
+}